requirements/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements/requirements.in
#
beautifulsoup4==4.13.3
    # via -r requirements/requirements.in
branca==0.8.1
    # via folium
certifi==2025.1.31
    # via
    #   pyogrio
    #   pyproj
    #   requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.1
    # via
    #   pdfminer-six
    #   requests
click==8.1.8
    # via -r requirements/requirements.in
cryptography==44.0.1
    # via pdfminer-six
folium==0.19.5
    # via -r requirements/requirements.in
fuzzywuzzy==0.18.0
    # via -r requirements/requirements.in
geopandas==1.0.1
    # via -r requirements/requirements.in
greenlet==3.1.1
    # via playwright
idna==3.10
    # via requests
jinja2==3.1.5
    # via
    #   branca
    #   folium
lxml==5.3.1
    # via python-docx
markupsafe==3.0.2
    # via jinja2
numpy==2.2.3
    # via
    #   folium
    #   geopandas
    #   pandas
    #   pyogrio
    #   shapely
packaging==24.2
    # via
    #   geopandas
    #   pyogrio
pandas==2.2.3
    # via
    #   -r requirements/requirements.in
    #   geopandas
pdfminer-six==20231228
    # via pdfplumber
pdfplumber==0.11.5
    # via -r requirements/requirements.in
pillow==11.1.0
    # via pdfplumber
playwright==1.50.0
    # via -r requirements/requirements.in
pycparser==2.22
    # via cffi
pyee==12.1.1
    # via playwright
pyogrio==0.10.0
    # via geopandas
pypdfium2==4.30.1
    # via pdfplumber
pyproj==3.7.1
    # via geopandas
python-dateutil==2.9.0.post0
    # via pandas
python-docx==1.1.2
    # via -r requirements/requirements.in
pytz==2025.1
    # via pandas
requests==2.32.3
    # via
    #   -r requirements/requirements.in
    #   folium
shapely==2.0.7
    # via geopandas
six==1.17.0
    # via python-dateutil
soupsieve==2.6
    # via beautifulsoup4
typing-extensions==4.12.2
    # via
    #   beautifulsoup4
    #   pyee
    #   python-docx
tzdata==2025.1
    # via pandas
urllib3==2.3.0
    # via requests
xyzservices==2025.1.0
    # via folium
